+++ docs/sphinx/changelog.rst
@@ -26,6 +26,7 @@ New Features
 - Added linear temperature to internal energy relation (:pull:`185`). By `Nathan Miller`_.
 - Added the ability to modify the global residuals and Jacobians (:pull:`186`). By `Nathan Miller`_.
 - Added a micromorphic linear elasticity model that is compatible with radial return algorithms (:pull:`187`). By `Nathan Miller`_.
+- Allow the user to do a complete build of the tardigrade stack (:pull:`188`). By `Nathan Miller`_.
 
 Internal Changes
 ================